refactor: add no data page on SFT page
